requirements.txt
ag2[gemini]
python-dotenv
praw